change lecture model
