+++ lecture/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from .models import LectureMpdel
 
 # Register your models here.
+admin.site.register(LectureMpdel)
