+++ upload/admin.py
@@ -1,5 +1,2 @@
 from django.contrib import admin
-from .models import uploadFileMpdel
 # Register your models here.
-
-admin.site.register(uploadFileMpdel)
